+++ PomodoroTimer/index.html
@@ -41,6 +41,6 @@
 
 </body>
 
-<script src="js/script.js"></script>
+<script src="js/script.js" type="module"></script>
 
 </html>
